Conflicts:
	lib/view/upload.html
	public/css/upload.css

+++ lib/utils/converter_checker/check_converter_by_file_content.php
@@ -115,6 +115,14 @@ class ConvertersMonitor {
 
     public function performCheck() {
 
+        $lockHandler = fopen( "$this->path/_flock.lock", 'w+' );
+        if( !flock( $lockHandler, LOCK_EX | LOCK_NB ) ){
+            self::_prettyEcho( "*************************************************************" );
+            self::_prettyEcho( "*********** Another Instance Running. SKIP CHECK. ***********" );
+            self::_prettyEcho( "*************************************************************" );
+            return;
+        }
+
         //clean
         #self::_prettyEcho( "Removing tmp files" );
         $this->_deleteDir( "$this->path/$this->converted_dir/" );
@@ -170,6 +178,10 @@ class ConvertersMonitor {
 
         $this->performRebooting();
 
+        fclose( $lockHandler );
+        unlink( "$this->path/_flock.lock" );
+
+        self::_prettyEcho( "----------- Released Lock ----------" );
         self::_prettyEcho( "------------------------------------" );
 
     }
@@ -287,10 +299,10 @@ class ConvertersMonitor {
             foreach ( $cmd as $c ) {
                 $stream = ssh2_exec( $conn, $c );
                 stream_set_blocking( $stream, true );
-                sleep(1);
                 $stream_out  = ssh2_fetch_stream( $stream, SSH2_STREAM_STDIO );
                 $ret_content = stream_get_contents( $stream_out );
-
+                sleep(15);
+                fclose( $stream );
                 $ret[ ] = "command is $c";
                 $ret[ ] = $ret_content;
 
@@ -490,8 +502,8 @@ class ConvertersMonitor {
             $rebootTime             = new DateTime( $this->resultSet[ $ip_converter ][ 'last_update' ] );
             $thisTimeFailure = new DateTime();
 
-            //if this failure happened 5 minutes after reboot time
-            if( $thisTimeFailure->modify('-5 minutes') >= $rebootTime ){
+            //if this failure happened 10 minutes after reboot time
+            if( $thisTimeFailure->modify('-10 minutes') >= $rebootTime ){
 
                 self::_prettyEcho( "> *** FAILED REBOOT FOUND....", 4 );
 
@@ -542,7 +554,7 @@ class ConvertersMonitor {
 //            }
 //
 //            if( $winword_total_load > 60 ){
-//                self::_prettyEcho( "> *** Found harmful instance of WINWORD, SET FOR REBOOT....", 4 );
+//                self::_prettyEcho( "> *** Found harmful instance, SET FOR REBOOT....", 4 );
 //                $this->setForReboot[ ] = $ip_converter;
 //            }
 //
